Add support for displayNameField setting for use in Dataverse/AROMA
In an Element a field may have isDisplayNameField=true/false value to
designate the field as the display name for the whole Element.

When an Element is used in a Template one can set a displayNameField value
in the element containing the name of a field in the Element to be the
designated display name for the whole Element. If one of the fields of the
Element already has an isDisplayNameField=true value, then this field will
be set as the displayNameField in the instance of the Element. This can be
later overwritten.

We allow selecting a value for displayNameField usign either a select input
listing the available field names, or using a text input where we may allow
setting a display field name expression not just a field name. Eg.

displayNameField=spaceConcat(name, title)

This would generate a display name by concatenating the 'name' and 'title'
fields using a space character. This is not supported yet, just a possible
feature in the future.