Move displayNameField setting to Element editor
Instead of allowing setting a field in Element editor and then redefining it
in the Template when it is instantiated, we only allow setting the
displayNameField value in the Element.

For now we also allow overriding it in the Template but only for admins
and will only use it to configure the builtin DV schemas, which would
otherwise be difficult.