Refact merge modal
The merge modal will still only be visible for admins and only if the template or element is a copy of its original resource.
Since, the "merge" functionality will only replace the ids of the updated resource with the ids of the original resource, this is the only case in which the merge modal can be used.