Add custom "dataverseFile" template element as input type - refs #24940
The custom input type can be added from the toolbar from the "More" section or of from the "Search" menu. In order to work from the toolbar, the custom template element must have the "schema:identifier" attribute set to "dataverseFile" and its id must be saved as an env variable that is used in the "url-service.conf.json" file.